impl: prompt user as per employee decision

+++ Unit4Ex3/src/main/java/Admin.java
@@ -3,7 +3,7 @@ import lombok.NonNull;
 public class Admin extends FullTimeTutor {
     private final double ADMIN_BOUNCE = 0.2;
 
-    public Admin(int staffId, String name, String address, int phoneNumber, boolean isPhDHolder, boolean isMasterHolder, double annualSalary) {
+    public Admin(long staffId, String name, String address, int phoneNumber, boolean isPhDHolder, boolean isMasterHolder, double annualSalary) {
         super(staffId, name, address, phoneNumber, isPhDHolder, isMasterHolder, annualSalary);
     }
 }
