+++ Unit4Ex3/src/main/java/SalariedTutor.java
@@ -2,7 +2,7 @@ import lombok.NonNull;
 
 public class SalariedTutor extends FullTimeTutor {
 
-    public SalariedTutor(int staffId, String name, String address, int phoneNumber, boolean isPhDHolder, boolean isMasterHolder, double annualSalary) {
+    public SalariedTutor(long staffId, String name, String address, int phoneNumber, boolean isPhDHolder, boolean isMasterHolder, double annualSalary) {
         super(staffId, name, address, phoneNumber, isPhDHolder, isMasterHolder, annualSalary);
     }
 }
